+++ walkers/src/download.rs
@@ -12,9 +12,7 @@ use image::ImageError;
 use reqwest::header::USER_AGENT;
 use reqwest_middleware::ClientWithMiddleware;
 
-use crate::{
-    http_tiles::HttpStats, io::http_client, mercator::TileId, sources::TileSource, tiles::Texture,
-};
+use crate::{http_tiles::HttpStats, io::http_client, sources::TileSource, tiles::Texture, TileId};
 
 pub use reqwest::header::HeaderValue;
