+++ walkers/src/lib.rs
@@ -16,7 +16,6 @@ mod zoom;
 pub use download::{HeaderValue, HttpOptions};
 pub use http_tiles::{HttpStats, HttpTiles};
 pub use map::{Map, MapMemory, Plugin, Projector};
-pub use mercator::TileId;
 pub use position::{lat_lon, lon_lat, Position};
-pub use tiles::{Texture, TextureWithUv, Tiles};
+pub use tiles::{Texture, TextureWithUv, TileId, Tiles};
 pub use zoom::InvalidZoom;
