+++ walkers/src/sources/mod.rs
@@ -4,7 +4,7 @@ mod geoportal;
 mod mapbox;
 mod openstreetmap;
 
-use crate::mercator::TileId;
+use crate::TileId;
 pub use geoportal::Geoportal;
 pub use mapbox::{Mapbox, MapboxStyle};
 pub use openstreetmap::OpenStreetMap;
